--- src/components/reservation/SectionDivider.tsx
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const SectionDivider = styled.div`
-  height: 8px;
-  background-color: ${(props) => props.theme.Color.borderColor};
-`;
